Return uint8_t from version getters

Fixes #149

+++ NK_C_API.h
@@ -573,13 +573,13 @@ extern "C" {
 	 * Get device's major firmware version
 	 * @return major part of the version number (e.g. 0 from 0.48, 0 from 0.7 etc.)
 	 */
-	NK_C_API int NK_get_major_firmware_version();
+	NK_C_API uint8_t NK_get_major_firmware_version();
 
 	/**
 	 * Get device's minor firmware version
 	 * @return minor part of the version number (e.g. 7 from 0.7, 48 from 0.48 etc.)
 	 */
-	NK_C_API int NK_get_minor_firmware_version();
+	NK_C_API uint8_t NK_get_minor_firmware_version();
 
   /**
    * Function to determine unencrypted volume PIN type
